--- version 2.0.0 (02-Sep-2016)---- adjustments to better rights management

+++ administrator/components/com_bwpostman/models/newsletter.php
@@ -115,63 +115,19 @@ class BwPostmanModelNewsletter extends JModelAdmin
 	}
 
 	/**
-	 * Method to test whether a record can be deleted.
-	 *
-	 * @param	object	$record	A record object.
-	 *
-	 * @return	boolean	True if allowed to delete the record. Defaults to the permission set in the component.
-	 *
-	 * @since	1.0.1
-	 */
-	protected function canDelete($record)
-	{
-		$user = JFactory::getUser();
-
-		// Check general delete permission first.
-		if ($user->authorise('bwpm.delete', 'com_bwpostman'))
-		{
-			return true;
-		}
-
-		if (!empty($record->id))
-		{
-			// Check specific delete permission.
-			if ($user->authorise('bwpm.newsletter.delete', 'com_bwpostman.newsletters.' . (int) $record->id))
-			{
-				return true;
-			}
-		}
-		return false;
-	}
-
-	/**
 	 * Method to test whether a record can have its state edited.
 	 *
 	 * @param	object	$record	A record object.
 	 *
-	 * @return	boolean	True if allowed to change the state of the record. Defaults to the permission set in the component.
+	 * @return	boolean	True if allowed to change the state of the record.
 	 *
 	 * @since	1.0.1
 	 */
 	protected function canEditState($record)
 	{
-		$user = JFactory::getUser();
+		$permission = BwPostmanHelper::canEditState('newsletter', $record->id);
 
-		// Check general edit state permission first.
-		if ($user->authorise('bwpm.edit.state', 'com_bwpostman'))
-		{
-			return true;
-		}
-
-		if (!empty($record->id))
-		{
-			// Check specific edit state permission.
-			if ($user->authorise('bwpm.newsletter.edit.state', 'com_bwpostman.newsletters.' . (int) $record->id))
-			{
-				return true;
-			}
-		}
-		return false;
+		return $permission;
 	}
 
 	/**
@@ -221,6 +177,14 @@ class BwPostmanModelNewsletter extends JModelAdmin
 				// Convert to the JObject before adding other data.
 				$properties = $table->getProperties(1);
 				$item       = ArrayHelper::toObject($properties, 'JObject');
+
+				// check permission
+				if (!BwPostmanHelper::canEdit('newsletter', $item))
+				{
+					// @ToDo: throw exception
+					$app->enqueueMessage(JText::_('COM_BWPOSTMAN_ERROR_EDIT_NO_PERMISSION'), 'error');
+					return false;
+				}
 
 				if (property_exists($item, 'params'))
 				{
@@ -942,31 +906,21 @@ class BwPostmanModelNewsletter extends JModelAdmin
 			$time = $date->toSql();
 
 			// Access check.
-			foreach ($cid as $i)
+			if (!BwPostmanHelper::canArchive('newsletter', $cid))
 			{
-				$data = self::getItem($i);
-				if (!BwPostmanHelper::allowArchive($i, $data->created_by, 'newsletter'))
-				{
-					$app->enqueueMessage(JText::_('COM_BWPOSTMAN_NL_ARCHIVE_RIGHTS_MISSING'), 'error');
-					return false;
-				}
+				return false;
 			}
 		}
 		else
 		{
+			// Access check.
+			if (!BwPostmanHelper::canRestore('newsletter', $cid))
+			{
+				return false;
+			}
+
 			$time	= '0000-00-00 00:00:00';
 			$uid	= 0;
-
-			// Access check.
-			foreach ($cid as $i)
-			{
-				$data = self::getItem($i);
-				if (!BwPostmanHelper::allowRestore($i, $data->created_by, 'newsletter'))
-				{
-					$app->enqueueMessage(JText::_('COM_BWPOSTMAN_NL_RESTORE_RIGHTS_MISSING'), 'error');
-					return false;
-				}
-			}
 		}
 
 		if (count($cid))
@@ -1005,6 +959,11 @@ class BwPostmanModelNewsletter extends JModelAdmin
 	 */
 	public function copy($cid = array())
 	{
+		if (!BwPostmanHelper::canAdd('newsletter'))
+		{
+			return false;
+		}
+
 		$app	= JFactory::getApplication();
 		$_db	= $this->_db;
 
@@ -1114,19 +1073,15 @@ class BwPostmanModelNewsletter extends JModelAdmin
 	 */
 	public function delete(&$pks)
 	{
-		// Access check.
-		foreach ($pks as $i)
-		{
-			$data = self::getItem($i);
-			if (!BwPostmanHelper::allowDelete($i, $data->created_by, 'newsletter'))
-			{
-				return false;
-			}
-		}
-
 		if (count($pks))
 		{
 			ArrayHelper::toInteger($pks);
+
+			// Access check.
+			if (!BwPostmanHelper::canDelete('newsletter', $pks))
+			{
+				return false;
+			}
 
 			// Delete newsletter from newsletters-table
 			$nl_table = JTable::getInstance('newsletters', 'BwPostmanTable');
@@ -1164,6 +1119,12 @@ class BwPostmanModelNewsletter extends JModelAdmin
 	 */
 	public function delete_queue()
 	{
+		// Access check
+		if (!BwPostmanHelper::canClearQueue())
+		{
+			return false;
+		}
+
 		$_db	= $this->_db;
 
 		$query = "TRUNCATE TABLE {$_db->quoteName('#__bwpostman_sendmailqueue')} ";
@@ -1686,6 +1647,12 @@ class BwPostmanModelNewsletter extends JModelAdmin
 	 */
 	public function sendNewsletter(&$ret_msg, $recipients, $nl_id, $unconfirmed, $cam_id)
 	{
+		// Access check
+		if (!BwPostmanHelper::canSend($nl_id))
+		{
+			return false;
+		}
+
 		// Prepare the newsletter content
 		$id	= $this->_addSendMailContent($nl_id);
 		if ($id	=== false)
@@ -1713,14 +1680,21 @@ class BwPostmanModelNewsletter extends JModelAdmin
 	/**
 	 * Method to reset the count of sending attempts in sendmailqueue.
 	 *
-	 * @return void
+	 * @return bool
 	 *
 	 * @since
 	 */
 	public function resetSendAttempts()
 	{
+		// Access check
+		if (!BwPostmanHelper::canResetQueue())
+		{
+			return false;
+		}
+
 		$tblSendmailQueue = $this->getTable('sendmailqueue', 'BwPostmanTable');
 		$tblSendmailQueue->resetTrials();
+		return true;
 	}
 
 	/**
